[IMP]将buy.py拆分成多个文件，及其他整合
