+++ buy/__openerp__.py
@@ -9,9 +9,9 @@
     'version': '8.0.0.1',
     'depends': ['core', 'warehouse', 'money', 'finance'],
     'data': [
-        'buy_data.xml',
+        'data/buy_data.xml',
         'security/groups.xml',
-        'buy_view.xml',
+        'views/buy_view.xml',
         'wizard/buy_order_track_wizard_view.xml',
         'wizard/buy_order_detail_wizard_view.xml',
         'wizard/buy_summary_goods_wizard_view.xml',
@@ -26,9 +26,9 @@
         'report/supplier_statements_view.xml',
         'report/report_data.xml',
         'security/ir.model.access.csv',
-        'home_page_data.xml'
+        'data/home_page_data.xml'
         ],
     'demo': [
-             'buy_demo.xml',
+             'data/buy_demo.xml',
              ],
 }
