+++ buy/models/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+import buy_order
+import buy_receipt
+import buy_adjust
+import money
+import vendor_goods
