+++ buy/views/buy_view.xml
